Client build: break up grunt tasks into separate files for clarity
